requirements.txt
python-pptx==0.6.23
pandas==1.5.3